#21: improved reporting

+++ monitoring/monitoring-analysis/analysis-report/src/main/java/com/rbmhtechnology/vind/monitoring/report/ReportService.java
@@ -95,7 +95,7 @@ public abstract class ReportService implements AutoCloseable {
     //Total number of requests
     public abstract long getTotalRequests();
     //Top Day
-    public abstract LinkedHashMap<ZonedDateTime, Integer> getTopDays();
+    public abstract LinkedHashMap<ZonedDateTime, Long> getTopDays();
 
     //Top users
     public abstract LinkedHashMap<String, Long> getTopUsers();
